pyproject.toml
[project]
name = "agentic-evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
]
